Move some test utils to main lib's tests
Move the UniformColorVideoSource.cs and LocalOnlySignaler.s files from
the samples package to the main library package, to be used in tests in
the latter too.
